Fix build error on Windows after BLI_read changes
Don't use posix-only ssize_t in platform independent code.

Pull Request: https://projects.blender.org/blender/blender/pulls/113493

+++ source/blender/blenlib/BLI_fileops.h
@@ -296,7 +296,7 @@ int BLI_access(const char *filepath, int mode) ATTR_WARN_UNUSED_RESULT ATTR_NONN
  * - Reads `size_t` bytes instead of `int` on WIN32.
  * \return the number of bytes read.
  */
-ssize_t BLI_read(int fd, void *buf, size_t nbytes);
+int64_t BLI_read(int fd, void *buf, size_t nbytes);
 
 /**
  * Returns true if the file with the specified name can be written.
