+++ source/blender/blenlib/intern/fileops.c
@@ -92,7 +92,7 @@ static char *windows_operation_string(FileExternalOperation operation)
 }
 #endif
 
-ssize_t BLI_read(int fd, void *buf, size_t nbytes)
+int64_t BLI_read(int fd, void *buf, size_t nbytes)
 {
   /* Define our own read as `read` is not guaranteed to read the number of bytes requested.
    * This happens rarely but was observed with larger than 2GB files on Linux, see: #113473.
@@ -100,9 +100,9 @@ ssize_t BLI_read(int fd, void *buf, size_t nbytes)
    * Even though this is a loop, the most common code-path will exit with "Success" case.
    * In the case where read more data than the file contains, it will loop twice,
    * exiting on EOF with the second iteration. */
-  ssize_t nbytes_read_total = 0;
+  int64_t nbytes_read_total = 0;
   while (true) {
-    ssize_t nbytes_read = read(fd,
+    int64_t nbytes_read = read(fd,
                                buf,
 #ifdef WIN32
                                /* Read must not exceed INT_MAX on WIN32, clamp. */
@@ -1239,7 +1239,7 @@ static int copy_single_file(const char *from, const char *to)
     /* symbolic links should be copied in special way */
     char *link_buffer;
     int need_free;
-    ssize_t link_len;
+    int64_t link_len;
 
     /* get large enough buffer to read link content */
     if ((st.st_size + 1) < sizeof(buf)) {
